Corrected code issues
