+++ src/hooks/useGuesses.ts
@@ -48,34 +48,48 @@ class GuessStorage {
   }
 }
 
-export function useGameResults(dayString: string) {
-  const [gameResultInStorage, setTodaysGameResultInStorageSetter] = useState<GameResult>(
-      GuessStorage.loadGameResultForDay(dayString)
-  );
+const calculateGameResult = (guesses: Guess[], maxGuesses: number) => {
+  console.log("Calculating game result with", guesses, maxGuesses)
+  const lastGuessIdx = guesses.length;
+  if(lastGuessIdx === 0) {
+    return "ONGOING"
+  }
+  
+  const lastGuess = guesses[lastGuessIdx-1];
+  if (lastGuess.distance === 0) {
+    return "VICTORY";
+  }
+  if (lastGuessIdx + 1 === maxGuesses) {
+    return "LOSS";
+  }
+  return "ONGOING";
+};
 
-  const setTodaysGameResultToLocalStorage = (gameRes: GameResult) => {
-    setTodaysGameResultInStorageSetter(gameRes);
-    GuessStorage.updateGameResults(dayString, gameRes)
-  };
-
-
-  return {gameResultInStorage: gameResultInStorage, setTodaysGameResultToLocalStorage}
-
-}
-
-export function useGuesses(dayString: string) {
+/** 
+ * NOTE: should only be used within useGame
+ * - usage multiple places may lead to bugs
+ */
+export function useGuesses(dayString: string, maxAttempts: number) {
   const [guesses, setGuesses] = useState<Guess[]>(
     GuessStorage.loadGuessesForDay(dayString)
+  );
+  const [gameResult, setGameResult] = useState<GameResult>(
+    GuessStorage.loadGameResultForDay(dayString)
   );
 
   const addGuess = (guess: Guess) => {
     const newGuesses = [...guesses, guess];
     setGuesses(newGuesses);
     GuessStorage.updateSavedGuesses(dayString, newGuesses);
+
+    const newGameResult = calculateGameResult(guesses, maxAttempts);
+    GuessStorage.updateGameResults(dayString, newGameResult);
+    setGameResult(newGameResult);
   };
 
   return {
     guesses,
     addGuess,
+    gameResult,
   };
 }
